+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# 17.12.2023 4:33AM
+- [feat]: Calling prescribeMedication
+- [feat]: Implemented history fetching
+
 # 17.12.2023 4:12AM
 - [feat]: Doctor getPatientVitals implemented
 - [feat]: checkin, checkout implemented for patient